Move charts text in public
