+++ packages/x-charts/src/ChartsText/index.tsx
@@ -1,0 +1,1 @@
+export { ChartsTextStyle, ChartsTextProps, ChartsText } from './ChartsText';
